Added pseudo-ORM; registration functional.
User registration is functional and records get added.

A pseudo-ORM class was made for sake of consistency. The issue
with using an ORM is that there maybe possible persistence
issues with it; defining a pseudo-ORM gives greater debug
capability and added flexibility.

The README has been updated.

Easter Egg 1: At the time of this commit, I am fully aware
test driven development is not being used. This is going to
a focus of the next sprint and those to come, as well as
build system. In the meantime, #YOLO.